requirements.txt
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
aiohttp>=3.8.0
openai>=1.0.0
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0
python-dotenv>=1.0.0